Refactor ui elements into ui module
