+++ app/build.gradle
@@ -116,4 +116,6 @@ dependencies {
 
     // Desugar JDK
     coreLibraryDesugaring 'com.android.tools:desugar_jdk_libs:1.2.0'
+
+    implementation(project(":core:ui"))
 }
