+++ app/src/main/java/com/xdiach/diarymoodapp/presentation/screens/authentication/AuthenticationScreenLayout.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.xdiach.diarymoodapp.R
-import com.xdiach.diarymoodapp.presentation.components.GoogleButton
+import com.xdiach.ui.components.GoogleButton
 import com.xdiach.diarymoodapp.ui.values.Dimensions
 
 @Composable
