+++ app/src/main/java/com/xdiach/diarymoodapp/presentation/screens/home/HomeScreenLayout.kt
@@ -26,7 +26,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.xdiach.diarymoodapp.R
 import com.xdiach.diarymoodapp.model.Diary
-import com.xdiach.diarymoodapp.presentation.components.DiaryHolder
+import com.xdiach.ui.components.DiaryHolder
 import com.xdiach.diarymoodapp.ui.values.Dimensions
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
