+++ app/src/main/java/com/xdiach/diarymoodapp/presentation/screens/write/WriteTopBar.kt
@@ -33,7 +33,7 @@ import com.maxkeppeler.sheets.clock.ClockDialog
 import com.maxkeppeler.sheets.clock.models.ClockSelection
 import com.xdiach.diarymoodapp.R
 import com.xdiach.diarymoodapp.model.Diary
-import com.xdiach.diarymoodapp.presentation.components.DisplayAlertDialog
+import com.xdiach.ui.components.DisplayAlertDialog
 import com.xdiach.diarymoodapp.ui.UiText
 import com.xdiach.diarymoodapp.util.toInstant
 import java.text.DateFormat
