+++ core/ui/src/main/java/com/xdiach/ui/components/AlertDialog.kt
@@ -1,4 +1,4 @@
-package com.xdiach.diarymoodapp.presentation.components
+package com.xdiach.ui.components
 
 import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
