+++ core/ui/src/main/java/com/xdiach/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.xdiach.diarymoodapp.ui.theme
+package com.xdiach.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
