+++ core/ui/src/main/res/drawable/google_logo.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+    <path
+        android:pathData="M21.752,12.224a11.544,11.544 0,0 0,-0.158 -1.885h-9.382v3.745h5.372a4.626,4.626 0,0 1,-1.993 2.972v2.491h3.2A9.742,9.742 0,0 0,21.752 12.224Z"
+        android:fillColor="#4285f4"/>
+    <path
+        android:pathData="M12.212,21.963a9.491,9.491 0,0 0,6.584 -2.416l-3.2,-2.491a6.011,6.011 0,0 1,-3.379 0.963,5.95 5.95,0 0,1 -5.588,-4.118h-3.3v2.566A9.949,9.949 0,0 0,12.212 21.963Z"
+        android:fillColor="#34a853"/>
+    <path
+        android:pathData="M6.624,13.901a5.781,5.781 0,0 1,-0.316 -1.9,6.007 6.007,0 0,1 0.316,-1.9L6.624,7.533h-3.3a9.848,9.848 0,0 0,-1.071 4.467,9.848 9.848,0 0,0 1.071,4.467Z"
+        android:fillColor="#fbbc05"/>
+    <path
+        android:pathData="M12.212,5.98a5.406,5.406 0,0 1,3.819 1.495l2.84,-2.84a9.526,9.526 0,0 0,-6.659 -2.6,9.949 9.949,0 0,0 -8.893,5.5l3.3,2.566A5.95,5.95 0,0 1,12.212 5.98Z"
+        android:fillColor="#ea4335"/>
+</vector>
